requirements.txt
streamlit==1.28.1
yt-dlp==2024.10.22
pydub==0.25.1
openai-whisper==20231117
librosa==0.10.1
numpy==1.24.3
scikit-learn==1.3.0
requests==2.32.4
openai==1.3.8
tenacity==8.2.3
certifi>=2025.6.15 